refactor(search): extract parallel provider search from hybrid engine
- Add hybrid-search-provider-parallel.ts with executeProviderSearchesWithOverallTimeout
- Slim HybridSearchEngine: runProviderSearchTaskBody, smaller createProviderSearchTask
- Move experiment_id logging to logRankingExperimentIfApplicable

Refs #315
